ci: unify changelog cap logic into a single script
Collapse generate-pr-changelog.js and prepare-changelog.js into
changelog.js with two CLI modes (pr / cap). The shared grapheme-cap
primitive (capGraphemes) is now defined once; both call sites are
updated to invoke the new entry point.
